+++ R/constants.R
@@ -1,5 +1,5 @@
 constants <- list(
-  INPUTS_FILENAME = "inputs.yaml",
+  CONFIGURATION_FILENAME = "project.yaml",
   LOCAL_INPUTS_FILENAME = "inputs.local.yaml",
   ANALYSES_DIR = "analyses",
   README_FILENAME = "README.md",